Adds error handling for create request
