+++ eth_crowdfund_be/eth_crowdfund_api/src/views/__init__.py
@@ -1,0 +1,8 @@
+from flask import Response, json
+
+def custom_response(res, status_code):
+  return Response(
+    mimetype="application/json",
+    response=json.dumps(res),
+    status=status_code
+  )
